% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glob_lasso_vis.R
\name{glob_lasso_vis}
\alias{glob_lasso_vis}
\title{Global Visualization of SHAP Values for a Lasso Regression Model}
\usage{
glob_lasso_vis(
  vip_featured,
  hiv_data,
  lasso_hyperparameters,
  vip_train,
  v_train
)
}
\arguments{
\item{vip_featured}{The name of the response variable to explain.}

\item{hiv_data}{The training dataset containing predictor variables and the response variable.}

\item{lasso_hyperparameters}{A list of hyperparameters for the Lasso model, including:
\itemize{
\item \code{penalty}: Regularization parameter
}}

\item{vip_train}{The dataset used for training the Lasso model.}

\item{v_train}{The response variable used for training the Lasso model.}
}
\value{
A visualization of global feature importance for the CR model.
}
\description{
This function generates a visualization for the global feature importance of
Lasso regression model trained on HIV data with specified
hyperparameters.
}
\examples{
\dontrun{
library(qvirus)
library(dplyr)

set.seed(123)

hiv_data <- qphen[, -1]

lasso_hyperparameters <- list(
  penalty = 0.01
)

vip_featured <- "cd_diff"

vip_train <- hiv_data |>
  dplyr::select(-all_of(vip_featured))

v_train <- hiv_data |>
  dplyr::select(all_of(vip_featured))

glob_lasso_vis(
  vip_featured = vip_featured,
  hiv_data = hiv_data,
  lasso_hyperparameters = lasso_hyperparameters,
  vip_train = vip_train,
  v_train = v_train
)
}
}
