% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train2-data.R
\docType{data}
\name{train2}
\alias{train2}
\title{Training Data for Explainability of Models}
\format{
A tibble (data frame) with 25 rows and 6 columns.
}
\usage{
data(train2)
}
\description{
This dataset contains training data for viral load explainer models. It
includes CD4 and viral load measurements for different years.
}
\note{
To explore more rows of this dataset, you can use the \code{print(n = ...)} function.
}
\examples{
data(train2)
train2
}
\author{
Juan Pablo Acuña González \href{mailto:22253567@uagro.mx}{22253567@uagro.mx}
}
\keyword{datasets}
